+++ components/d2l-activity-editor/d2l-activity-assignment-editor/d2l-activity-assignment-editor-detail.js
@@ -1,17 +1,15 @@
 import 'd2l-inputs/d2l-input-text.js';
 import { css, html, LitElement } from 'lit-element/lit-element.js';
+import { ACELocalizeMixin } from './d2l-activity-assignment-editor-localize-mixin';
 import { AssignmentEntity } from 'siren-sdk/src/activities/assignments/AssignmentEntity.js';
 import { Debouncer } from '@polymer/polymer/lib/utils/debounce.js';
 import { EntityMixinLit } from 'siren-sdk/src/mixin/entity-mixin-lit.js';
 import { ErrorHandlingMixin } from '../error-handling-mixin.js';
 import { labelStyles } from '@brightspace-ui/core/components/typography/styles.js';
-import { LocalizeMixin } from '@brightspace-ui/core/mixins/localize-mixin.js';
 import { SirenFetchMixinLit } from 'siren-sdk/src/mixin/siren-fetch-mixin-lit.js';
 import { timeOut } from '@polymer/polymer/lib/utils/async.js';
 
-const SUPPORTED_LANGUAGES = ['en', 'fr'];
-
-class AssignmentEditorDetail extends ErrorHandlingMixin(SirenFetchMixinLit(EntityMixinLit(LocalizeMixin(LitElement)))) {
+class AssignmentEditorDetail extends ErrorHandlingMixin(SirenFetchMixinLit(EntityMixinLit(ACELocalizeMixin(LitElement)))) {
 
 	static get properties() {
 		return {
@@ -30,28 +28,6 @@ class AssignmentEditorDetail extends ErrorHandlingMixin(SirenFetchMixinLit(Entit
 				display: none;
 			}
 		`];
-	}
-
-	static async getLocalizeResources(langs) {
-		const supportedLanguages = langs.reverse().filter(language => {
-			return SUPPORTED_LANGUAGES.indexOf(language) > -1;
-		});
-		const requests = supportedLanguages.map(language => {
-			return fetch(`./lang/${language}.json`).then(res => res.json());
-		});
-		const responses = await Promise.all(requests);
-
-		const langterms = {};
-		responses.forEach(language => {
-			for (const langterm in language) {
-				langterms[langterm] = language[langterm].translation;
-			}
-		});
-
-		return {
-			language: supportedLanguages[supportedLanguages.length - 1],
-			resources: langterms
-		};
 	}
 
 	constructor() {
