+++ demo/d2l-activity-editor/d2l-activity-assignment-editor/d2l-activity-assignment-editor-detail.html
@@ -14,7 +14,7 @@
 	</head>
 	<body>
 		<script>
-			function changeLocale(locale) {
+			function changeLocale(locale) { // eslint-disable-line no-unused-vars
 				document.getElementsByTagName('html')[0].lang = locale;
 			}
 		</script>
