--- demo/d2l-activity-editor/d2l-activity-assignment-editor/lang/en.json
@@ -1,10 +1,0 @@
-{
-  "name": {
-    "translation": "Name",
-    "context": "Helpful hints to make the Serge translation better"
-  },
-  "emptyNameError": {
-    "translation": "Name is required",
-    "context": "Helpful hints to make the Serge translation better"
-  }
-}
