--- demo/d2l-activity-editor/d2l-activity-assignment-editor/lang/fr.json
@@ -1,10 +1,0 @@
-{
-  "name": {
-    "translation": "Nom",
-    "context": "Helpful hints to make the Serge translation better"
-  },
-  "emptyNameError": {
-    "translation": "Le nom est requis",
-    "context": "Helpful hints to make the Serge translation better"
-  }
-}
